+++ packages/matter-node-shell.js/src/app.ts
@@ -1,17 +1,7 @@
 /**
- * Copyright 2022 Project CHIP Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * @license
+ * Copyright 2022-2023 Project CHIP Authors
+ * SPDX-License-Identifier: Apache-2.0
  */
 
 import { BleNode } from "@project-chip/matter-node-ble.js/ble";
@@ -90,8 +80,9 @@ async function main() {
 
                 const { nodeNum, ble, nodeType, factoryReset, netInterface } = argv;
 
-                const theNode = new MatterNode(nodeNum, netInterface);
+                theNode = new MatterNode(nodeNum, netInterface);
                 await theNode.initialize(factoryReset);
+
                 const theShell = new Shell(theNode, PROMPT);
 
                 if (ble) {