+++ packages/matter-node-shell.js/src/shell/cmd_nodes.ts
@@ -1,21 +1,52 @@
 /**
- * Copyright 2022 Project CHIP Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * @license
+ * Copyright 2022-2023 Project CHIP Authors
+ * SPDX-License-Identifier: Apache-2.0
  */
 
+import { NodeId } from "@project-chip/matter-node.js/datatype";
+import { CommissioningControllerNodeOptions, NodeStateInformation } from "@project-chip/matter-node.js/device";
+import { Logger } from "@project-chip/matter-node.js/log";
 import type { Argv } from "yargs";
 import { MatterNode } from "../MatterNode";
+
+export function createDiagnosticCallbacks(): Partial<CommissioningControllerNodeOptions> {
+    return {
+        attributeChangedCallback: (peerNodeId, { path: { nodeId, clusterId, endpointId, attributeName }, value }) =>
+            console.log(
+                `attributeChangedCallback ${peerNodeId}: Attribute ${nodeId}/${endpointId}/${clusterId}/${attributeName} changed to ${Logger.toJSON(
+                    value,
+                )}`,
+            ),
+        eventTriggeredCallback: (peerNodeId, { path: { nodeId, clusterId, endpointId, eventName }, events }) =>
+            console.log(
+                `eventTriggeredCallback ${peerNodeId}: Event ${nodeId}/${endpointId}/${clusterId}/${eventName} triggered with ${Logger.toJSON(
+                    events,
+                )}`,
+            ),
+        stateInformationCallback: (peerNodeId, info) => {
+            switch (info) {
+                case NodeStateInformation.Connected:
+                    console.log(`stateInformationCallback Node ${peerNodeId} connected`);
+                    break;
+                case NodeStateInformation.Disconnected:
+                    console.log(`stateInformationCallback Node ${peerNodeId} disconnected`);
+                    break;
+                case NodeStateInformation.Reconnecting:
+                    console.log(`stateInformationCallback Node ${peerNodeId} reconnecting`);
+                    break;
+                case NodeStateInformation.WaitingForDeviceDiscovery:
+                    console.log(
+                        `stateInformationCallback Node ${peerNodeId} waiting that device gets discovered again`,
+                    );
+                    break;
+                case NodeStateInformation.StructureChanged:
+                    console.log(`stateInformationCallback Node ${peerNodeId} structure changed`);
+                    break;
+            }
+        },
+    };
+}
 
 export default function commands(theNode: MatterNode) {
     return {
@@ -72,7 +103,95 @@ export default function commands(theNode: MatterNode) {
                         const node = (await theNode.connectAndGetNodes(nodeId))[0];
 
                         console.log("Logging structure of Node ", node.nodeId.toString());
-                        node.logStructure();
+                        node.logStructure({});
+                    },
+                )
+                .command(
+                    "connect <node-id> [min-subscription-interval] [max-subscription-interval]",
+                    "Connects to one or all cmmissioned nodes",
+                    yargs => {
+                        return yargs
+                            .positional("node-id", {
+                                describe: "node id to connect. Use 'all' to connect to all nodes.",
+                                default: undefined,
+                                type: "string",
+                                demandOption: true,
+                            })
+                            .positional("min-subscription-interval", {
+                                describe:
+                                    "Minimum subscription interval in seconds. If set then the node is subscribed to all attributes and events.",
+                                type: "number",
+                            })
+                            .positional("max-subscription-interval", {
+                                describe:
+                                    "Maximum subscription interval in seconds. If minimum interval is set and this not this is set to 30 seconds.",
+                                type: "number",
+                            });
+                    },
+                    async argv => {
+                        const { nodeId: nodeIdStr, maxSubscriptionInterval, minSubscriptionInterval } = argv;
+                        await theNode.start();
+                        if (theNode.commissioningController === undefined) {
+                            throw new Error("CommissioningController not initialized");
+                        }
+                        let nodeIds = theNode.commissioningController.getCommissionedNodes();
+                        if (nodeIdStr !== "all") {
+                            const cmdNodeId = NodeId(BigInt(nodeIdStr));
+                            nodeIds = nodeIds.filter(nodeId => nodeId === cmdNodeId);
+                            if (!nodeIds.length) {
+                                throw new Error(`Node ${nodeIdStr} not commissioned`);
+                            }
+                        }
+
+                        const autoSubscribe = minSubscriptionInterval !== undefined;
+
+                        for (const nodeIdToProcess of nodeIds) {
+                            await theNode.commissioningController.connectNode(nodeIdToProcess, {
+                                autoSubscribe,
+                                subscribeMinIntervalFloorSeconds: autoSubscribe ? minSubscriptionInterval : undefined,
+                                subscribeMaxIntervalCeilingSeconds: autoSubscribe
+                                    ? maxSubscriptionInterval ?? 30
+                                    : undefined,
+                                ...createDiagnosticCallbacks(),
+                            });
+                        }
+                    },
+                )
+                .command(
+                    "disconnect <node-id>",
+                    "Disconnects from one or all nodes",
+                    yargs => {
+                        return yargs.positional("node-id", {
+                            describe: "node id to disconnect. Use 'all' to disconnect from all nodes.",
+                            default: undefined,
+                            type: "string",
+                            demandOption: true,
+                        });
+                    },
+                    async argv => {
+                        const { nodeId: nodeIdStr } = argv;
+                        if (theNode.commissioningController === undefined) {
+                            console.log("Controller not initialized, nothing to disconnect.");
+                            return;
+                        }
+
+                        let nodeIds = theNode.commissioningController.getCommissionedNodes();
+                        if (nodeIdStr !== "all") {
+                            const cmdNodeId = NodeId(BigInt(nodeIdStr));
+                            nodeIds = nodeIds.filter(nodeId => nodeId === cmdNodeId);
+                            if (!nodeIds.length) {
+                                throw new Error(`Node ${nodeIdStr} not commissioned`);
+                            }
+                        }
+
+                        for (const nodeIdToProcess of nodeIds) {
+                            const node = theNode.commissioningController.getConnectedNode(nodeIdToProcess);
+                            if (node === undefined) {
+                                console.log(`Node ${nodeIdToProcess} not connected`);
+                                continue;
+                            }
+                            await node.disconnect();
+                        }
                     },
                 ),
         handler: async (argv: any) => {